More ports from 2.8.
git-svn-id: http://chianti.ucsd.edu/svn@22876 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ core3/ding-presentation-impl/trunk/src/main/java/org/cytoscape/ding/ArrowShape.java
@@ -84,7 +84,7 @@ public enum ArrowShape {
 	;
 
 
-	private static Map<Integer,Shape> arrowShapes = GraphGraphics.getArrowShapes();
+	private static Map<Byte,Shape> arrowShapes = GraphGraphics.getArrowShapes();
 
 	private String shapeName;
 	private String ginyShapeName;
